@@ -11,35 +11,39 @@
     "prettier-write": "prettier --write --ignore-unknown"
   },
   "dependencies": {
-    "@angular/animations": "13.3.11",
+    "@angular/animations": "14.2.0",
     "@angular/cdk": "13.3.9",
-    "@angular/common": "13.3.11",
-    "@angular/compiler": "13.3.11",
-    "@angular/core": "13.3.11",
-    "@angular/forms": "13.3.11",
+    "@angular/common": "14.2.0",
+    "@angular/compiler": "14.2.0",
+    "@angular/core": "14.2.0",
+    "@angular/forms": "14.2.0",
     "@angular/material": "13.3.9",
-    "@angular/platform-browser": "13.3.11",
-    "@angular/platform-browser-dynamic": "13.3.11",
-    "@angular/router": "13.3.11",
+    "@angular/platform-browser": "14.2.0",
+    "@angular/platform-browser-dynamic": "14.2.0",
+    "@angular/router": "14.2.0",
+    "@ngx-translate/core": "^14.0.0",
+    "@ngx-translate/http-loader": "^7.0.0",
     "rxjs": "^6.6.6",
+    "three": "^0.143.0",
     "tslib": "^2.0.0",
     "zone.js": "0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.3.7",
+    "@angular-devkit/build-angular": "14.2.1",
     "@angular-eslint/builder": "13.2.1",
     "@angular-eslint/eslint-plugin": "13.2.1",
     "@angular-eslint/eslint-plugin-template": "13.2.1",
     "@angular-eslint/template-parser": "13.2.1",
-    "@angular/cli": "13.3.7",
-    "@angular/compiler-cli": "13.3.11",
-    "@angular/language-service": "13.3.11",
-    "@ngrx/effects": "13.0.2",
-    "@ngrx/store": "13.0.2",
-    "@ngrx/store-devtools": "13.0.2",
+    "@angular/cli": "14.2.1",
+    "@angular/compiler-cli": "14.2.0",
+    "@angular/language-service": "14.2.0",
+    "@ngrx/effects": "14.3.0",
+    "@ngrx/store": "14.3.0",
+    "@ngrx/store-devtools": "14.3.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
+    "@types/three": "^0.143.1",
     "@typescript-eslint/eslint-plugin": "5.18.0",
     "@typescript-eslint/parser": "5.18.0",
     "eslint": "8.12.0",
