@@ -1,7 +1,9 @@
-import { LayoutModule as AngularLayoutModule } from '@angular/cdk/layout';
+
 import { CommonModule, Location, LocationStrategy, PathLocationStrategy } from '@angular/common';
-import { HttpClientJsonpModule, HttpClientModule } from '@angular/common/http';
+import {  HttpClientModule } from '@angular/common/http';
 import { NgModule, Optional, SkipSelf } from '@angular/core';
+
+// ngrx store
 import { AppStoreModule } from '@store/store.module';
 import { LayoutModule } from './layout/layout.module';
 import { moduleImportGuard } from './module-import-guard';
@@ -13,10 +15,8 @@ import { RoutingModule } from './routing/routing.module';
         RoutingModule
     ],
     imports: [
-        AngularLayoutModule,
         AppStoreModule,
         CommonModule,
-        HttpClientJsonpModule,
         HttpClientModule,
         LayoutModule,
         RoutingModule
