@@ -1,20 +1,46 @@
+// angular
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { HttpClient } from '@angular/common/http';
+import { AppComponent } from './app.component';
+
+
 import { CoreModule } from '@core/core.module';
 import { FeaturesModule } from '@features/features.module';
-import { AppComponent } from './app.component';
+
+
+// i18n
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { ThreejsBackgroundComponent } from '@core/layout/threejs-background/threejs-background.component';
+
+export function HttpLoaderFactory(http: HttpClient) {
+    return new TranslateHttpLoader(http,'content/i18n/');
+  }
+
 
 @NgModule({
     bootstrap: [
         AppComponent
     ],
     declarations: [
-        AppComponent
+        AppComponent,
+        ThreejsBackgroundComponent
     ],
     imports: [
         BrowserModule,
         CoreModule,
-        FeaturesModule
+        FeaturesModule,
+        TranslateModule.forRoot({
+            defaultLanguage: 'en',
+            loader: {
+              provide: TranslateLoader,
+              useFactory: HttpLoaderFactory,
+              deps: [HttpClient]
+            }
+          })        
+        
+        
     ]
 })
 export class AppModule { }
