@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { LayoutModule } from '@core/layout/layout.module';
+import { TranslateModule } from '@ngx-translate/core';
 import { HomePageComponent } from './home-page';
 import { ProfileModule } from './profile/profile.module';
 
@@ -12,7 +13,8 @@ import { ProfileModule } from './profile/profile.module';
     ],
     imports: [
         LayoutModule,
-        ProfileModule
+        ProfileModule,
+        TranslateModule
     ]
 })
 export class FeaturesModule { }
