@@ -1,16 +1,36 @@
+// angular
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { MatCardModule } from '@angular/material/card';
-import { MatDividerModule } from '@angular/material/divider';
-import { MatListModule } from '@angular/material/list';
+import { HttpClientModule } from '@angular/common/http';
+import { RouterModule } from '@angular/router';
+
+
+// miss
 import { LayoutModule } from '@core/layout/layout.module';
+import { SharedModule } from 'src/app/shared/shared.module';
+
+// @ngrx
 import { StoreModule } from '@ngrx/store';
-import { ProfileDetailComponent } from './profile-detail';
 import { getProfileReducer } from './store/profile.reducers';
+import { EffectsModule } from '@ngrx/effects';
+import { ProfileEffects } from './store/profile.effects';
+
+// i18n
+import { TranslateModule } from '@ngx-translate/core';
+
+// profile components
+import { ProfileListComponent } from './profile-list/profile-list.component';
+import { ProfileMainComponent } from './profile-main/profile-main.component';
+import { ProfileDetailComponent } from './profile-detail';
+import { ProfileDetailCardModule } from 'projects/profile-detail-card/src/public-api';
+
 
 @NgModule({
     declarations: [
-        ProfileDetailComponent
+        ProfileDetailComponent,
+        ProfileListComponent,
+        ProfileMainComponent,
+
     ],
     exports: [
         ProfileDetailComponent
@@ -18,10 +38,13 @@ import { getProfileReducer } from './store/profile.reducers';
     imports: [
         CommonModule,
         LayoutModule,
-        MatCardModule,
-        MatDividerModule,
-        MatListModule,
-        StoreModule.forFeature('profile', getProfileReducer)
+        SharedModule,
+        ProfileDetailCardModule,
+        HttpClientModule,
+        TranslateModule,
+        RouterModule,
+        StoreModule.forFeature('profile', getProfileReducer),
+        EffectsModule.forFeature([ProfileEffects])
     ]
 })
 export class ProfileModule { }
