@@ -1,29 +1,24 @@
 import { ProfileState } from '@interfaces';
 import { Action, createReducer, on } from '@ngrx/store';
-import { profileActions } from '@store/actions';
-import { UserProfile } from '../interfaces';
+import { ProfileActions } from './index';
 
-const dummyProfile: UserProfile = {
-    cellNumber: '888-888-8888',
-    city: 'Los Angeles',
-    dateOfBirth: 'Jan 1st, 1966',
-    email: 'test@crexi.com',
-    firstName: 'First Name',
-    lastName: 'Last Name',
-    phoneNumber: '999-999-9999',
-    picture: '/content/img/default_user.png',
-    state: 'CA'
-};
 
-const initialState: ProfileState = {};
+const initialState: ProfileState = new ProfileState();
 
 const reducer = createReducer(
-    initialState,
-    on(profileActions.initProfile, (state) => ({ ...state, user: dummyProfile }))
+  initialState,
+  on(ProfileActions.listRandomProfileSuccess, (state, action) => 
+  ({ ...state, users: [...state.users,...action.result] })
+  ),
+  on(ProfileActions.updateCurrentUserId, (state, action) => {
+    
+    return ({ ...state, currentUserId: action.id })
+  }
+  )
 );
 
-export function getProfileReducer (state: ProfileState | undefined, action: Action) {
+export function getProfileReducer(state: ProfileState | undefined, action: Action) {
 
-    return reducer(state, action);
+  return reducer(state, action);
 
 }
